requirements.txt
# ================================
# Core Web Framework
# ================================
fastapi==0.128.0
uvicorn==0.40.0

# ================================
# Data Validation & Parsing
# ================================
pydantic==2.12.3
pydantic-settings==2.12.0
jsonschema==4.25.1

# ================================
# Environment Variables
# ================================
python==3.13.7

# ================================
# LLM / Generative AI (Bonus)
# ================================
google-generativeai==0.8.6

# ================================
# Testing
# ================================
pytest==9.0.2

# ================================
# Utilities
# ================================
requests==2.32.5
